refactor: change project structure according to FSD architecture
