+++ tsconfig.app.json
@@ -7,6 +7,16 @@
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "module": "ESNext",
     "skipLibCheck": true,
+    "baseUrl": "./src",
+    "paths": {
+      "@app/*": ["./app/*"],
+      "@processes/*": ["./processes/*"],
+      "@pages/*": ["./pages/*"],
+      "@widgets/*": ["./widgets/*"],
+      "@features/*": ["./features/*"],
+      "@entities/*": ["./entities/*"],
+      "@shared/*": ["./shared/*"]
+    },
 
     /* Bundler mode */
     "moduleResolution": "bundler",
